requirements.txt
pyyaml
Cython
loguru
h5py
einops
pyclipper
pycocotools
Shapely
timm
segmentation-models-pytorch
torchsummary
imgaug
opencv-python
